Report cleanups, initialize_columns -> add_sheet refactor
remove unused (and inaccurate) LOG_PREFIX from a bunch of places
remove copypased (and inaccurate) comments
fix a couple of comments

move _build_data_section_usage_by_node_with_org_details & _build_data_section_usage_by_job from report base to report_ccsp_v2

replace _init_dimensions for main sheet + same for loop in each _build_* function,
move into base add_sheet(title, sheet_index, columns=None)

Fail with "No billing data" only when all dataframes are empty
([0] may be None if no sheets are using it even when data is available)
