+++ metrics_utility/automation_controller_billing/dataframe_engine/dataframe_inventory_scope.py
@@ -15,8 +15,6 @@ logger = logging.getLogger(__name__)
 
 
 class DataframeInventoryScope(Base):
-    LOG_PREFIX = '[AAPBillingReport] '
-
     def build_dataframe(self):
         # A daily rollup dataframe
