+++ metrics_utility/automation_controller_billing/dataframe_engine/dataframe_jobhost_summary_usage.py
@@ -16,8 +16,6 @@ logger = logging.getLogger(__name__)
 
 
 class DataframeJobhostSummaryUsage(Base):
-    LOG_PREFIX = '[AAPBillingReport] '
-
     def build_dataframe(self):
         # A daily rollup dataframe
         billing_data_monthly_rollup = None
