+++ metrics_utility/automation_controller_billing/report_saver/report_saver_directory.py
@@ -3,8 +3,6 @@ import os
 
 
 class ReportSaverDirectory:
-    LOG_PREFIX = '[ExtractorDirectory]'
-
     def __init__(self, extra_params, logger=logging.getLogger(__name__)):
         self.extra_params = extra_params
